pyproject.toml
[project]
name = "finding-donors-project"
version = "0.1.0"
description = "Apply the Supervised Learning techniques learned from Udacity lectures"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "joblib>=1.5.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "pytest>=9.0.3",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
